+++ pkg/client/cmd/login_cluster.go
@@ -36,7 +36,7 @@ func login(cmd *cobra.Command, args []string) {
 		client.PrintErrorAndExit("Error trying to get the user password: %v", err)
 	}
 
-	conn, err := connection.New(cfgFile)
+	conn, err := connection.New(cfgFile, cfgCluster)
 	if err != nil {
 		client.PrintErrorAndExit("Error connecting to server: %v", err)
 	}